created modal/dialog for adding tasks
