+++ src/components/TaskDialog.jsx
@@ -1,0 +1,66 @@
+/* eslint-disable object-curly-newline */
+import React, { useState } from 'react'
+import { Dialog, Grid, Button, TextField, DialogContent, DialogActions, DialogTitle } from '@material-ui/core'
+import useAddTask from '../graphql/task/hooks/useAddTask'
+
+const TaskDialog = ({ dialogStatus, column, toggleDialog }) => {
+    const [addTask] = useAddTask()
+    const [title, setTitle] = useState('')
+
+    const handleChange = (event) => {
+        setTitle(event.target.value)
+    }
+
+    const handleSave = (event) => {
+        event.preventDefault()
+        addTask({
+            variables: {
+                columnId: column.id,
+                title,
+            },
+        })
+        toggleDialog()
+    }
+
+    return (
+        <Grid>
+            <Dialog
+                fullWidth
+                maxWidth="md"
+                onClose={toggleDialog}
+                open={dialogStatus}
+                aria-labelledby="max-width-dialog-title"
+            >
+                <DialogTitle aria-labelledby="max-width-dialog-title">Create new task</DialogTitle>
+                <DialogContent>
+                    <TextField
+                        autoComplete="off"
+                        margin="dense"
+                        name="title"
+                        label="Name"
+                        type="text"
+                        value={title}
+                        fullWidth
+                        onChange={handleChange}
+                    />
+                </DialogContent>
+                <DialogActions>
+                    <Button
+                        onClick={toggleDialog}
+                        color="secondary"
+                    >
+                        Cancel
+                    </Button>
+                    <Button
+                        disabled={!title.length}
+                        onClick={handleSave}
+                        color="primary"
+                    >
+                        Create task
+                    </Button>
+                </DialogActions>
+            </Dialog>
+        </Grid>
+    )
+}
+export default TaskDialog
